Rewrite tests for `BaseAlertViewSet` as unit tests
