+++ tests/unit/metrics/api/views/test_alerts.py
@@ -1,4 +1,22 @@
+from unittest import mock
+from django.urls import reverse
+
+from django.test import RequestFactory
+
 from metrics.api.views.alerts import BaseAlertViewSet
+import pytest
+
+
+class InvalidMetricExtendedBaseAlertViewSet(BaseAlertViewSet):
+    @property
+    def topic_name(self) -> str:
+        return "topic_name"
+
+
+class InvalidTopicExtendedBaseAlertViewSet(BaseAlertViewSet):
+    @property
+    def metric_name(self) -> str:
+        return "metric_name"
 
 
 class TestBaseAlertsView:
@@ -16,3 +34,35 @@ class TestBaseAlertsView:
 
         # Then
         assert permission_classess == []
+
+    def test_raises_error_if_topic_name_not_implemented(self):
+        """
+        Given an instance of the `BaseAlertViewSet`
+        When the `topic_name` is not implemented in the child class
+        Then a `NotImplementedError` is raised.
+        """
+        # Given
+        path = reverse("cold-alerts-list")
+        fake_request = RequestFactory().get(path)
+        fake_request.query_params = mock.MagicMock()
+        extended_base_alert_view_set = InvalidTopicExtendedBaseAlertViewSet()
+
+        # When / Then
+        with pytest.raises(NotImplementedError):
+            extended_base_alert_view_set.list(fake_request)
+
+    def test_raises_error_if_metric_name_not_implemented(self):
+        """
+        Given an instance of the `BaseAlertViewSet`
+        When the `metric_name` is not implemented in the child class
+        Then a `NotImplementedError` is raised.
+        """
+        # Given
+        path = reverse("cold-alerts-list")
+        fake_request = RequestFactory().get(path)
+        fake_request.query_params = mock.MagicMock()
+        extended_base_alert_view_set = InvalidMetricExtendedBaseAlertViewSet()
+
+        # When / Then
+        with pytest.raises(NotImplementedError):
+            extended_base_alert_view_set.list(fake_request)
